% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmobs-package.R
\docType{package}
\name{msmobs-package}
\alias{msmobs}
\alias{msmobs-package}
\title{msmobs: Fit Marginal Structural Models for Observational Data}
\description{
Estimate weights and fit pooled logistic regression outcome models for marginal structural models in observational data analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zabore/msmobs}
  \item Report bugs at \url{https://github.com/zabore/msmobs/issues}
}

}
\author{
\strong{Maintainer}: Emily Zabor \email{emily.zabor@gmail.com}

Authors:
\itemize{
  \item Emily Zabor \email{emily.zabor@gmail.com}
}

}
\keyword{internal}
